Refactored most code
