+++ src/components/header/MobileHeader.tsx
@@ -1,0 +1,51 @@
+import { Dialog, DialogPanel } from '@headlessui/react'
+import { XMarkIcon } from '@heroicons/react/24/outline'
+import { Iprops } from './headerTypes'
+
+export const MobileHeader = ({
+  mobileMenuOpen,
+  setMobileMenuOpen,
+  navigation
+}: Iprops) => {
+  return (
+    <Dialog
+      open={mobileMenuOpen}
+      onClose={setMobileMenuOpen}
+      className="lg:hidden"
+    >
+      <div className="fixed inset-0 z-50" />
+      <DialogPanel className="fixed inset-y-0 right-0 z-50 w-full overflow-y-auto bg-white p-6 sm:max-w-sm sm:ring-1 sm:ring-gray-900/10">
+        <div className="flex items-center justify-between">
+          <a href="#" className="-m-1.5 p-1.5">
+            <span className="h-8 w-auto text-2xl font-bold text-black">
+              Kunal
+            </span>
+          </a>
+          <button
+            type="button"
+            onClick={() => setMobileMenuOpen(false)}
+            className="-m-2.5 rounded-md p-2.5 text-gray-700"
+          >
+            <span className="sr-only">Close menu</span>
+            <XMarkIcon aria-hidden="true" className="size-6" />
+          </button>
+        </div>
+        <div className="mt-6 flow-root">
+          <div className="-my-6 divide-y">
+            <div className="space-y-2 py-6">
+              {navigation.map((item) => (
+                <a
+                  key={item.name}
+                  href={item.href}
+                  className="-mx-3 block rounded-lg px-3 py-2 text-base/7 font-semibold text-gray-900 hover:bg-gray-50"
+                >
+                  {item.name}
+                </a>
+              ))}
+            </div>
+          </div>
+        </div>
+      </DialogPanel>
+    </Dialog>
+  )
+}
