+++ src/components/header/headerTypes.ts
@@ -1,0 +1,10 @@
+type NavItem = {
+  name: string
+  href: string
+}
+
+export type Iprops = {
+  mobileMenuOpen?: boolean
+  setMobileMenuOpen: React.Dispatch<React.SetStateAction<boolean>>
+  navigation: NavItem[]
+}
